Remove JavaScript files after no longer needed
This apparently won't save memory as all the functions
are still in memory if the HTML tag to the JS file is removed.
Maybe the garbage collector will clear up some of them?
Anyway, the main benefit is debuggability, as you can see which JS files
are currently intended to be used